cleanup reportBrowserCrash
